+++ src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment: any = {
     production: true,
+    staging: false,
     hmr: false
 };