+++ src/environments/environment.ts
@@ -5,5 +5,6 @@
 
 export const environment: any = {
     production: false,
+    staging: false,
     hmr: false
 };